+++ src/Interpreter/Interpreter.Expression.cs
@@ -8,8 +8,6 @@
  Licence: MIT
 */
 
-
-
 using BazzBasic.Lexer;
 using BazzBasic.Parser;