+++ src/Interpreter/Interpreter.Math.cs
@@ -10,7 +10,7 @@
 
 using BazzBasic.Lexer;
 using BazzBasic.Parser;
-using System.Reflection.Metadata.Ecma335;
+//using System.Reflection.Metadata.Ecma335;
 
 namespace BazzBasic.Interpreter;